Reorganize file functions package
