--- core/src/main/scala/com/eharmony/spotz/util/CrossValidation.scala
@@ -1,8 +1,0 @@
-package com.eharmony.spotz.util
-
-/**
-  * @author vsuthichai
-  */
-trait CrossValidation {
-
-}
